Additional tweaks

+++ llm_functions.py
@@ -1,3 +1,4 @@
+import json
 import pathlib
 import shlex
 import subprocess
@@ -32,8 +33,64 @@ def _initialized_db_connection(path: pathlib.Path) -> sqlite3.Connection:
     return connection
 
 
-def fresh_db_connection() -> sqlite3.Connection:
+def fresh_db_connection(*, overwrite=True) -> sqlite3.Connection:
     output_db = pathlib.Path("./chatbot.db")
-    if output_db.exists():
+    if overwrite and output_db.exists():
         output_db.unlink()
     return _initialized_db_connection(output_db)
+
+
+def create_tables(connection: sqlite3.Connection):
+    vector_size = len(embedding_for_text("Dummy string to get back an embedding"))
+    with connection as transaction:
+        transaction.execute(
+            f"""
+                create virtual table vec_chat using vec0(
+                    embedding float[{vector_size}]
+                );
+            """
+        )
+
+        transaction.execute(
+            """
+                create table text_lines(
+                    filename text,
+                    text_line text
+                );
+            """
+        )
+
+
+def nearest_n_neighbors(
+    text: str, connection: sqlite3.Connection, limit: int = 5
+) -> list[int]:
+    with connection as transaction:
+        return [
+            x
+            for x in transaction.execute(
+                """
+                    select rowid, distance
+                    from vec_chat
+                    where sample_embedding match ?
+                    order by distance
+                    limit ?;
+                """,
+                json.dumps(embedding_for_text(text)),
+                limit,
+            )
+        ]
+
+
+def nearest_n_text_lines(
+    text: str, connection: sqlite3.Connection, limit: int = 5
+) -> list[str]:
+    with connection as transaction:
+        return [
+            row[0]
+            for row in connection.execute(
+                """
+                    select text_line from text_lines where rowid in ? 
+                """,
+                nearest_n_neighbors(text, connection, limit),
+            )
+        ]
